requirements_cs.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements_cs.in --python-version 3.8
black==22.10.0
    # via -r requirements_cs.in
cerebras-appliance==2.5.0
    # via
    #   -r requirements_cs.in
    #   cerebras-sdk
cerebras-sdk==2.5.0
    # via -r requirements_cs.in
click==8.1.8
    # via black
dill==0.4.0
    # via cerebras-appliance
exceptiongroup==1.3.0
    # via pytest
grpcio==1.47.0
    # via
    #   cerebras-appliance
    #   grpcio-tools
grpcio-tools==1.47.0
    # via cerebras-appliance
h5py==3.10.0
    # via
    #   cerebras-appliance
    #   hdf5plugin
hdf5plugin==5.1.0
    # via cerebras-appliance
iniconfig==2.1.0
    # via pytest
mypy-extensions==1.1.0
    # via black
numpy==1.24.4
    # via
    #   cerebras-appliance
    #   h5py
packaging==25.0
    # via
    #   cerebras-appliance
    #   pytest
pathspec==0.12.1
    # via black
pip==25.0.1
    # via cerebras-appliance
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
protobuf==3.15.6
    # via
    #   cerebras-appliance
    #   grpcio-tools
psutil==7.1.0
    # via cerebras-appliance
pyjwt==2.9.0
    # via cerebras-appliance
pytest==8.3.5
    # via -r requirements_cs.in
pyyaml==6.0.3
    # via cerebras-appliance
ruff==0.14.0
    # via -r requirements_cs.in
setuptools==75.3.2
    # via grpcio-tools
six==1.17.0
    # via
    #   grpcio
    #   protobuf
tblib==1.7.0
    # via cerebras-appliance
tomli==2.3.0
    # via
    #   black
    #   pytest
tqdm==4.67.1
    # via cerebras-appliance
typing-extensions==4.13.2
    # via
    #   black
    #   exceptiongroup
urllib3==1.26.20
    # via cerebras-appliance
zstandard==0.23.0
    # via cerebras-appliance
